+++ app/src/main/java/com/jonas/cookly/core/util/sideeffect/SideEffect.kt
@@ -1,4 +1,4 @@
-package com.jonas.cookly.core.sideeffects
+package com.jonas.cookly.core.util.sideeffect
 
 sealed interface SideEffect {
     data class ShowToast(val message: String) : SideEffect
