+++ app/src/main/java/com/jonas/cookly/core/util/sideeffect/SingleEventEffect.kt
@@ -1,4 +1,4 @@
-package com.jonas.cookly.core.util
+package com.jonas.cookly.core.util.sideeffect
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
